+++ src/actions/composite.js
@@ -5,6 +5,7 @@ import { fetchNibrsCounts } from '../actions/nibrsCounts'
 import { fetchSummarized } from '../actions/summarized'
 import { fetchSummaries } from '../actions/summary'
 import { fetchLeoka } from '../actions/leoka'
+import { fetchFootnote } from '../actions/footnote'
 import { fetchUcrParticipation } from '../actions/participation'
 // import { fetchAgencies } from '../actions/agencies'
 import history, { createNewLocation } from '../util/history'
@@ -29,7 +30,11 @@ const fetchData = () => (dispatch, getState) => {
         dispatch(fetchSummaries(filters, states))
       }
       dispatch(fetchPoliceEmployment(filters))
-      if (filters.placeType === 'agency') dispatch(fetchSummarized(filters))
+      if (filters.placeType === 'agency') {
+        dispatch(fetchSummarized(filters))
+        dispatch(fetchSummaries(filters, states))
+        dispatch(fetchFootnote(filters))
+      }
       if (shouldFetchNibrs(filters, states)) dispatch(fetchNibrsCounts(filters))
     }
     if (filters.page === 'leoka') { // Add validation of leoka type and add a shouldFetch Method