+++ src/components/UsaMap.js
@@ -5,7 +5,6 @@ import Hint from './Hint'
 import stateLookup from '../util/usa'
 import svgData from '../../public/data/usa-state-svg.json'
 
-
 import { lookupStateByAbbr, lookupStateByName, lookupRegionByCode, lookupStatesByRegion } from '../util/location'
 
 class UsaMap extends React.Component {